fix: add Suspense boundaries for cacheComponents build
Enable cacheComponents: true in Next.js config and add Suspense
boundaries to all pages that access dynamic/runtime data so the
production build succeeds with partial prerendering.

Changes:
- Remove dynamicIO: false from next.config.js (legacy flag)
- Move auth from root layout to client-side (AuthProvider fetches
  user in useEffect instead of server-side getUser)
- Wrap AppBar in Suspense in (main) layout with static skeleton
- Extract page content into Suspense-wrapped async server components
  for: threads list, thread detail, CLI auth, auth login, profile,
  and embed pages
- Create featured threads data loader with connection() to defer
  to request time inside Suspense
- Add skeleton fallback components for streaming UI
