+++ app/src/app/(main)/layout.tsx
@@ -1,4 +1,7 @@
+import { Suspense } from "react";
+
 import { AppBar } from "@/components/ui/app-bar";
+import { AppBarSkeleton } from "@/components/ui/app-bar-skeleton";
 import { Footer } from "@/components/ui/footer";
 import { Toaster } from "@/components/ui/sonner";
 
@@ -9,7 +12,9 @@ type MainLayoutProps = {
 const MainLayout = ({ children }: MainLayoutProps) => {
   return (
     <>
-      <AppBar />
+      <Suspense fallback={<AppBarSkeleton />}>
+        <AppBar />
+      </Suspense>
       <main>{children}</main>
       <Footer />
       <Toaster />
