+++ app/src/app/(main)/threads/page.tsx
@@ -1,14 +1,13 @@
+import { Suspense } from "react";
 import { getTranslations } from "next-intl/server";
-
-import { createClient } from "@/server/supabase/server";
-import { sessions } from "@/server/repos/sessions";
 
 import { gradient } from "@/utils/renderers";
 
 import { Container } from "@/components/ui/container";
 import { Typography } from "@/components/ui/typography";
 
-import { ThreadsPageThreadsContainer } from "@/containers/threads-page-threads-container";
+import { ThreadsPageDataLoader } from "@/components/threads-page-data-loader";
+import { ThreadsPageSkeleton } from "@/components/threads-page-skeleton";
 
 type ThreadsPageProps = {
   searchParams: Promise<{ query?: string }>;
@@ -16,10 +15,6 @@ type ThreadsPageProps = {
 
 const ThreadsPage = async ({ searchParams }: ThreadsPageProps) => {
   const t = await getTranslations();
-  const supabase = await createClient();
-
-  const { query } = await searchParams;
-  const { threads, total, nextOffset } = await sessions.getPublicThreads(supabase, { query });
 
   return (
     <Container size="md" spacing="md" className="flex flex-col gap-8">
@@ -32,12 +27,9 @@ const ThreadsPage = async ({ searchParams }: ThreadsPageProps) => {
         </Typography>
       </div>
 
-      <ThreadsPageThreadsContainer
-        initialQuery={query}
-        initialTotal={total}
-        initialNextOffset={nextOffset}
-        initialThreads={threads}
-      />
+      <Suspense fallback={<ThreadsPageSkeleton />}>
+        <ThreadsPageDataLoader searchParams={searchParams} />
+      </Suspense>
     </Container>
   );
 };
